SMS bugs & require phone/email/address on check in

+++ classes/Member/Address/Emails.php
@@ -3,7 +3,10 @@
 namespace HHK\Member\Address;
 
 use HHK\AuditLog\NameLog;
+use HHK\Exception\RuntimeException;
+use HHK\Exception\ValidationException;
 use HHK\HTMLControls\{HTMLContainer, HTMLInput, HTMLTable};
+use HHK\SysConst\EmailPurpose;
 use HHK\Tables\EditRS;
 use HHK\Tables\Name\NameEmailRS;
 
@@ -121,7 +124,7 @@ class Emails extends AbstractContactPoint {
 
         $adrRS = $this->get_recordSet($code);
 
-        if (is_null($adrRS) || $adrRS->Email->getStoredVal() == '') {
+        if (is_null($adrRS) || ($code !== EmailPurpose::NoEmail && $adrRS->Email->getStoredVal() == '')) {
             return FALSE;
         } else {
             return TRUE;
@@ -190,17 +193,22 @@ class Emails extends AbstractContactPoint {
                 .' ' .HTMLInput::generateMarkup($p[0], $prefAttr)
                 , array('class'=>"tdlabel " .$p[2]));
 
+        if ($p[0] == EmailPurpose::NoEmail) {
+            $tdContents = '';
+        } else {
+            // email address
+            $attr = array();
+            $attr['id'] = $idPrefix . 'txtEmail' . $p[0];
+            $attr['name'] = $idPrefix . 'txtEmail[' . $p[0] . ']';
+            $attr['title'] = 'Enter an email address';
+            $attr['class'] = 'hhk-emailInput ' . $inputClass;
+            $attr['size'] = '26';
 
-        // email address
-        $attr = array();
-        $attr['id'] = $idPrefix.'txtEmail' . $p[0];
-        $attr['name'] = $idPrefix.'txtEmail[' . $p[0] . ']';
-        $attr['title'] = 'Enter an email address';
-        $attr['class'] = 'hhk-emailInput ' . $inputClass;
-        $attr['size'] = '26';
+            $tdContents = HTMLInput::generateMarkup($this->rSs[$p[0]]->Email->getStoredVal(), $attr);
+        }
 
         //add input td
-        $trContents .= HTMLContainer::generateMarkup('td', HTMLInput::generateMarkup($this->rSs[$p[0]]->Email->getStoredVal(), $attr), array('class' => $p[2]));
+        $trContents .= HTMLContainer::generateMarkup('td', $tdContents, array('class' => $p[2]));
         
         return $trContents;
     }
@@ -238,7 +246,7 @@ class Emails extends AbstractContactPoint {
                 if ($a->idName->getStoredVal() > 0) {
                     // Email Address exists in the DB
 
-                    if ($post[$idPrefix.'txtEmail'][$purpose[0]] == '') {
+                    if ($post[$idPrefix.'txtEmail'][$purpose[0]] == '' && $purpose[0] !== EmailPurpose::NoEmail) {
 
                         // Delete the Email Address record
                         if (EditRS::delete($dbh, $a, array($a->idName, $a->Purpose)) === FALSE) {
@@ -263,7 +271,7 @@ class Emails extends AbstractContactPoint {
                 } else {
                     // Email Address does not exist inthe DB.
                     // Did the user fill in this Email Address panel?
-                    if ($post[$idPrefix.'txtEmail'][$purpose[0]] != '') {
+                    if ($post[$idPrefix.'txtEmail'][$purpose[0]] != ''|| $purpose[0] === EmailPurpose::NoEmail) {
 
                         // Insert a new Email Address
                         $this->loadPostData($a, $post, $purpose[0], $user, $idPrefix);
@@ -298,8 +306,13 @@ class Emails extends AbstractContactPoint {
      * @return void
      */
     private function loadPostData(NameEmailRS $a, array $p, $typeCode, $uname, $idPrefix = "") {
+        $email = filter_var(filter_var($p[$idPrefix . 'txtEmail'][$typeCode], FILTER_SANITIZE_EMAIL), FILTER_VALIDATE_EMAIL);
 
-        $a->Email->setNewVal(strtolower(trim(filter_var($p[$idPrefix.'txtEmail'][$typeCode], FILTER_SANITIZE_EMAIL))));
+        if($email == false){
+            throw new ValidationException("Email field must be a valid Email address");
+        }
+
+        $a->Email->setNewVal(strtolower(trim($email)));
         $a->Status->setNewVal("a");
         $a->Updated_By->setNewVal($uname);
         $a->Last_Updated->setNewVal(date("Y-m-d H:i:s"));
